fix(gui): solid sliding pill replaces the liquid glass idiom
The selection pill is now ONE persistent view that adopts
the selected segment's matched-geometry anchor — glass
conditionally attached to the selected label crossfaded
instead of sliding. Liquid Glass itself is dropped after
three variants (bare, tinted, clear + rim): it read as
washed-out, blue, or blurred the label. Solid white pill
(light gray in dark mode, no shadow), a real font-size step
for the selected label (scaleEffect rasterizes and blurs),
native white slider thumb overhanging a full-strength accent
fill, and a grayed fill when disabled (opacity alone let the
accent tint the knob blue). Types renamed to
SegmentedPicker / SettingsSlider — nothing glass remains.
